server/requirements.txt
py-quantaq==1.3.0
pandas==2.3.3
psycopg2-binary==2.9.11
fastapi==0.115.4
uvicorn==0.32.0
